pyproject.toml
[project]
name = "sun18-may-26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "gtts>=2.5.4",
    "openai-agents>=0.0.15",
    "python-dotenv>=1.1.0",
]
